pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pytelegrambotapi>=4.29.1",
    "python-dotenv>=1.2.1",
]
